// frontend/src/scrapers/kauflandScrapper/kauflandScrapper.types.ts
export interface KauflandProductData {
  productFound: boolean;
  kauflandProductId?: string | null;
  productName?: string | null;
  kauflandPrice?: string | null;
  kauflandLink?: string | null;
  kauflandSellerFee?: string | null;
  kauflandVat?: string | null;
  kauflandVariableFee?: string | null;
  kauflandFixedFee?: string | null;
  kauflandShippingRate?: string | null;
}